-Changed timepassed to u32. -Changed GetTicks function to u32.

+++ PSP Project/Include/EndingScreen.h
@@ -20,7 +20,7 @@ public:
 
 	virtual void draw();
 
-	virtual void update(u64 timePassed);
+	virtual void update(u32 timePassed);
 
 };
 #endif
