+++ PSP Project/Include/ExitScreen.h
@@ -18,6 +18,6 @@ public:
 
 	virtual void draw();
 
-	virtual void update(u64 timePassed);
+	virtual void update(u32 timePassed);
 };
 #endif
