+++ PSP Project/Include/Map.h
@@ -39,7 +39,7 @@ public:
 	bool EndOfOneWave(); //Same as EndOfWave but only returns true once
 	void EnemyKilled(); // Every time when an dies or reaches the end of the path.
 	void GetCurrentWaveEnemy(string &enemyName, int &EnemyLevel);
-	bool SpawnUpdate(u64 timePassed);
+	bool SpawnUpdate(u32 timePassed);
 	~Wave();
 
 	string mPath;
